pyproject.toml
[project]
name = "my-mcp-servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[tool]>=1.6.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.6.0",
]
